Update patch set 1
Patch Set 1:

thanks. i don't own the device. i was under the assumption that the the flag modifies only the recovery code and that the recovery and kernel are built together and monolithic, as is the case of the 4210 devices.

in any case, i can upload a patchset that in addition enables NO_SECURE_DISCARD, would you approve of that?

take a look:

https://github.com/CyanogenMod/android_system_extras/blob/cm-11.0/ext4_utils/wipe.c

Patch-set: 1
Reviewer: Lanchon <9776@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0